* master: (182 commits)
  libclamav/qsort.c: fix CMP1 macro (bb#1769)
  handle floating chars
  libclamav/readdb.c: make sure static sigs with floating chars go into AC
  sigtool: --decode-sigs: handle alternatives
  sigtool: --decode-sigs: handle .ldb sigs
  inflateinit spam
  sigtool: --decode-sigs: handle .ndb sigs
  fix sig printing
  Set limit to 255 so that warning is shown if maxthreads*maxrec would exceed it.
  Enable more than 256 FD support on Solaris (bb #1764).
  fix memleaks
  sigtool: --decode-sigs; decode .db entries (bb#1246)
  fsk sxs, gimme back my dll hell!
  Fix Solaris build: cli_hex2ui has to be added to libclamav.map
  Don't error on unused functions.
  sigtool: basic sig decoding
  sigtool/sigtool.c: handle .ign2 files (bb#1625)
  libclamav/qsort.c: don't call med3 when using internal cmp
  win32: fix globbing
  unify DUPMAX def
  ...

Conflicts:
	.gitignore
	Makefile.am
	configure
	libclamav/Makefile.in
	libclamav/clamav.h
	libclamav/disasm.c
	libclamav/libclamav.map
	libclamav/matcher.c
	libclamav/others.c
	libclamav/pe.c
	libclamav/pe.h
	libclamav/readdb.c

+++ clamd/thrmgr.c
@@ -237,11 +237,11 @@ int thrmgr_printstats(int f)
 		print_queue(f, pool->single_queue, &tv_now);
 		mdprintf(f, "\n");
 		for(task = pool->tasks; task; task = task->nxt) {
-			long delta;
+			double delta;
 			size_t used, total;
 
 			delta = tv_now.tv_usec - task->tv.tv_usec;
-			delta += (tv_now.tv_sec - task->tv.tv_sec)*1000000;
+			delta += (tv_now.tv_sec - task->tv.tv_sec)*1000000.0;
 			mdprintf(f,"\t%s %f %s\n",
 					task->command ? task->command : "N/A",
 					delta/1e6,
